# server/docs/swagger.yaml
info:
  contact: {}
paths:
  /music/upsert:
    post:
      consumes:
      - application/json
      description: 获取用户详细信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 插入或更新music
      tags:
      - 用户
  /relay/{id}:
    delete:
      consumes:
      - application/json
      description: 获取用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 根据id删除relay
      tags:
      - 用户
  /relay/password/{pwd}:
    get:
      consumes:
      - application/json
      description: 获取用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: pwd
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 根据password获取relay
      tags:
      - 用户
  /relay/pin/{id}:
    get:
      consumes:
      - application/json
      description: 获取用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 固定某个relay
      tags:
      - 用户
  /relay/type/{type}:
    get:
      consumes:
      - application/json
      description: 获取用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: type
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 根据relayType获取relay列表
      tags:
      - 用户
  /relay/upsert:
    post:
      consumes:
      - application/json
      description: 获取用户详细信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 插入或更新relay
      tags:
      - 用户
swagger: "2.0"